update: split presets to folder, add const to utils, truncate user name in preset, upgrade deps
