+++ .eslintrc.cjs
@@ -2,7 +2,7 @@ module.exports = {
 	extends: './node_modules/@companion-module/tools/eslint/main.cjs',
 	overrides: [
 		{
-			files: ['*.ts'],
+			files: ['src/**/*.ts'],
 			rules: {
 				"no-use-before-define": "off",
 				"@typescript-eslint/no-use-before-define": "off",
