+++ package.json
@@ -12,29 +12,32 @@
 		"build": "rimraf dist && yarn build:main",
 		"build:main": "tsc -p tsconfig.build.json",
 		"build:watch": "tsc -p tsconfig.build.json --watch",
-		"lint": "eslint ./src --ext .ts",
+		"lint": "eslint ./src/** --ext .ts",
 		"postinstall": "yarn build"
 	},
 	"license": "MIT",
 	"prettier": "@companion-module/tools/.prettierrc.json",
 	"dependencies": {
-		"@companion-module/base": "^1.1.1",
+		"@companion-module/base": "~1.4",
 		"osc": "^2.4.3"
 	},
 	"devDependencies": {
-		"@companion-module/tools": "^0.5.2",
-		"@types/node": "^17.0.12",
-		"husky": "^7.0.4",
-		"lint-staged": "^12.3.2",
-		"rimraf": "^3.0.2",
+		"@companion-module/tools": "1.2.0",
+		"@types/node": "^18.0.0",
+		"husky": "^8.0.3",
+		"lint-staged": "^13.2.3",
+		"rimraf": "^5.0.1",
 		"typescript": "^4.5.5"
 	},
 	"lint-staged": {
-		"*.{js,json,md}": [
+		"*.{js,json,md,ts}": [
 			"prettier --write"
 		],
 		"*.ts": [
 			"yarn lint --fix"
 		]
+	},
+	"engines": {
+		"node": ">=18.0.0"
 	}
 }
